feat(code refactor) - Spliting seconding resources
